add firmware validation client function
