+++ pkg/api/v1/conditions/client/client.go
@@ -92,6 +92,11 @@ func (c *Client) ServerBiosControl(ctx context.Context,
 	return c.post(ctx, path, params)
 }
 
+func (c *Client) ValidateFirmwareSet(ctx context.Context,
+	params *v1types.FirmwareValidationRequest) (*v1types.ServerResponse, error) {
+	return c.post(ctx, "validateFirmware", params)
+}
+
 func (c *Client) ServerConditionCreate(ctx context.Context, serverID uuid.UUID, conditionKind rctypes.Kind, conditionCreate v1types.ConditionCreate) (*v1types.ServerResponse, error) {
 	path := fmt.Sprintf("servers/%s/condition/%s", serverID.String(), conditionKind)
